semaforo-bot/requirements.txt
# Dependencias del SemáforoBot
# Python 3.11+

# Exchange y trading (SIN CoinGlass - usamos CCXT directamente)
ccxt==4.2.0

# Análisis de datos
pandas==2.1.4
numpy==1.26.2
ta==0.11.0

# API y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Base de datos y memoria
redis==5.0.1
aioredis==2.0.1

# HTTP requests (para webhooks opcionales)
requests==2.31.0
aiohttp==3.9.1

# Utilidades
python-dateutil==2.8.2
pytz==2023.3